requirements.txt
flask==2.3.3
flask-cors==3.0.10
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.1.4
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
transformers==4.37.2
requests==2.31.0
chardet==5.2.0
werkzeug==3.0.1
